+++ src/app/page.tsx
@@ -1,5 +1,11 @@
+import { Stack } from '@/components/stack'
+
 function Home() {
-	return <main className='flex min-h-screen flex-col items-center '></main>
+	return (
+		<main className='flex min-h-screen flex-col items-center '>
+			<Stack />
+		</main>
+	)
 }
 
 export default Home
