+++ src/components/stack/StackPopup.tsx
